refactor: add id field for using it as int64

+++ internal/ent/client.go
@@ -269,7 +269,7 @@ func (c *UserClient) UpdateOne(u *User) *UserUpdateOne {
 }
 
 // UpdateOneID returns an update builder for the given id.
-func (c *UserClient) UpdateOneID(id int) *UserUpdateOne {
+func (c *UserClient) UpdateOneID(id int64) *UserUpdateOne {
 	mutation := newUserMutation(c.config, OpUpdateOne, withUserID(id))
 	return &UserUpdateOne{config: c.config, hooks: c.Hooks(), mutation: mutation}
 }
@@ -286,7 +286,7 @@ func (c *UserClient) DeleteOne(u *User) *UserDeleteOne {
 }
 
 // DeleteOneID returns a builder for deleting the given entity by its id.
-func (c *UserClient) DeleteOneID(id int) *UserDeleteOne {
+func (c *UserClient) DeleteOneID(id int64) *UserDeleteOne {
 	builder := c.Delete().Where(user.ID(id))
 	builder.mutation.id = &id
 	builder.mutation.op = OpDeleteOne
@@ -303,12 +303,12 @@ func (c *UserClient) Query() *UserQuery {
 }
 
 // Get returns a User entity by its id.
-func (c *UserClient) Get(ctx context.Context, id int) (*User, error) {
+func (c *UserClient) Get(ctx context.Context, id int64) (*User, error) {
 	return c.Query().Where(user.ID(id)).Only(ctx)
 }
 
 // GetX is like Get, but panics if an error occurs.
-func (c *UserClient) GetX(ctx context.Context, id int) *User {
+func (c *UserClient) GetX(ctx context.Context, id int64) *User {
 	obj, err := c.Get(ctx, id)
 	if err != nil {
 		panic(err)
@@ -402,7 +402,7 @@ func (c *UserOauthProviderClient) UpdateOne(uop *UserOauthProvider) *UserOauthPr
 }
 
 // UpdateOneID returns an update builder for the given id.
-func (c *UserOauthProviderClient) UpdateOneID(id int) *UserOauthProviderUpdateOne {
+func (c *UserOauthProviderClient) UpdateOneID(id int64) *UserOauthProviderUpdateOne {
 	mutation := newUserOauthProviderMutation(c.config, OpUpdateOne, withUserOauthProviderID(id))
 	return &UserOauthProviderUpdateOne{config: c.config, hooks: c.Hooks(), mutation: mutation}
 }
@@ -419,7 +419,7 @@ func (c *UserOauthProviderClient) DeleteOne(uop *UserOauthProvider) *UserOauthPr
 }
 
 // DeleteOneID returns a builder for deleting the given entity by its id.
-func (c *UserOauthProviderClient) DeleteOneID(id int) *UserOauthProviderDeleteOne {
+func (c *UserOauthProviderClient) DeleteOneID(id int64) *UserOauthProviderDeleteOne {
 	builder := c.Delete().Where(useroauthprovider.ID(id))
 	builder.mutation.id = &id
 	builder.mutation.op = OpDeleteOne
@@ -436,12 +436,12 @@ func (c *UserOauthProviderClient) Query() *UserOauthProviderQuery {
 }
 
 // Get returns a UserOauthProvider entity by its id.
-func (c *UserOauthProviderClient) Get(ctx context.Context, id int) (*UserOauthProvider, error) {
+func (c *UserOauthProviderClient) Get(ctx context.Context, id int64) (*UserOauthProvider, error) {
 	return c.Query().Where(useroauthprovider.ID(id)).Only(ctx)
 }
 
 // GetX is like Get, but panics if an error occurs.
-func (c *UserOauthProviderClient) GetX(ctx context.Context, id int) *UserOauthProvider {
+func (c *UserOauthProviderClient) GetX(ctx context.Context, id int64) *UserOauthProvider {
 	obj, err := c.Get(ctx, id)
 	if err != nil {
 		panic(err)
